+++ apps/cowswap-frontend/src/common/state/totalSurplusState/index.ts
@@ -1,2 +1,3 @@
-export { TotalSurplusUpdater } from './updaters'
-export { useTriggerTotalSurplusUpdateCallback, useTotalSurplus } from './hooks'
+export { TotalSurplusUpdater, invalidateSurplusSwr } from './updaters'
+export { SurplusInvalidationListenerUpdater } from './SurplusInvalidationListenerUpdater'
+export { useTotalSurplus } from './hooks'