+++ apps/cowswap-frontend/src/modules/application/containers/App/Updaters.tsx
@@ -37,7 +37,7 @@ import { TradeType, useTradeTypeInfo } from 'modules/trade'
 import { UsdPricesUpdater } from 'modules/usdAmount'
 import { LpTokensWithBalancesUpdater, PoolsInfoUpdater, VampireAttackUpdater } from 'modules/yield/shared'
 
-import { TotalSurplusUpdater } from 'common/state/totalSurplusState'
+import { SurplusInvalidationListenerUpdater } from 'common/state/totalSurplusState'
 import { AnnouncementsUpdater } from 'common/updaters/AnnouncementsUpdater'
 import { ConnectionStatusUpdater } from 'common/updaters/ConnectionStatusUpdater'
 import { FeatureFlagsUpdater } from 'common/updaters/FeatureFlagsUpdater'
@@ -96,7 +96,6 @@ export function Updaters(): ReactNode {
       <SpotPricesUpdater />
       <InjectedWidgetUpdater />
       <CowEventsUpdater />
-      <TotalSurplusUpdater />
       <UsdPricesUpdater />
       <OrdersNotificationsUpdater />
       <OrderProgressStateUpdater />
@@ -104,6 +103,7 @@ export function Updaters(): ReactNode {
       <OrderProgressEventsUpdater />
       <SolversInfoUpdater />
       <AnnouncementsUpdater />
+      <SurplusInvalidationListenerUpdater />
       <BridgingEnabledUpdater />
       <FaviconAnimationUpdater />