+++ apps/cowswap-frontend/src/utils/orderUtils/isPermitValidForOrder.ts
@@ -1,5 +1,4 @@
-import { areAddressesEqual } from '@cowprotocol/common-utils'
-import { COW_PROTOCOL_VAULT_RELAYER_ADDRESS, SupportedChainId } from '@cowprotocol/cow-sdk'
+import { COW_PROTOCOL_VAULT_RELAYER_ADDRESS, SupportedChainId, areAddressesEqual } from '@cowprotocol/cow-sdk'
 import { Erc20__factory, type Erc20Interface } from '@cowprotocol/cowswap-abis'
 import { oneInchPermitUtilsConsts } from '@cowprotocol/permit-utils'
 import { Interface } from '@ethersproject/abi'